creating StepLabel component and moving classNames to constants.ts
