+++ packages/stepper/src/constants.ts
@@ -1,0 +1,4 @@
+import { createUniqueClassName } from '@leafygreen-ui/lib';
+
+export const stepLabelClassName = createUniqueClassName();
+export const stepIconClassName = createUniqueClassName();
